requirements.txt
opencv-python>=4.9.0
PySide6>=6.6.0
PyOpenGL>=3.1.7
